src/requirements.txt
#pip3 freeze > requirements.txt
colorzero
gpiozero
pigpio
RPi.GPIO
pydantic
PyYAML
pywebio
chevron
paho-mqtt
pigpio
pillow
